requirements.txt
streamlit_lightweight_charts
pandas
numpy
streamlit
google-generativeai 